+++ training_benchmark.py
@@ -6,15 +6,10 @@ import torch.nn.functional as F
 from tqdm import tqdm
 
 from utils import emit_itt, get_dataset, get_model
+from utils.microbatch_reddit_quiver_gloo import SAGE
 from torch_geometric.loader import NeighborLoader
 from torch_geometric.nn import PNAConv
 from torch_geometric.profile import rename_profile_file, timeit, torch_profile
-
-supported_sets = {
-    'ogbn-mag': ['rgat', 'rgcn'],
-    'ogbn-products': ['edge_cnn', 'gat', 'gcn', 'pna', 'sage'],
-    'Reddit': ['edge_cnn', 'gat', 'gcn', 'pna', 'sage'],
-}
 
 
 def train_homo(model, loader, optimizer, device, progress_bar=True,
@@ -37,9 +32,7 @@ def train_homo(model, loader, optimizer, device, progress_bar=True,
         optimizer.step()
 
 
-
-def run(args: argparse.ArgumentParser) -> None:
-
+def main(args: argparse.ArgumentParser) -> None:
     device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
     # If we use a custom number of steps, then we need to use RandomSampler,
     # which already does shuffle.
@@ -47,111 +40,74 @@ def run(args: argparse.ArgumentParser) -> None:
 
     print('BENCHMARK STARTS')
     for dataset_name in args.datasets:
-        assert dataset_name in supported_sets.keys(
-        ), f"Dataset {dataset_name} isn't supported."
         print(f'Dataset: {dataset_name}')
         data, num_classes = get_dataset(dataset_name, args.root,
                                         args.use_sparse_tensor, args.bf16)
-        hetero = True if dataset_name == 'ogbn-mag' else False
-        mask = ('paper', data['paper'].train_mask
-                ) if dataset_name == 'ogbn-mag' else data.train_mask
-        degree = None
         if torch.cuda.is_available():
             amp = torch.cuda.amp.autocast(enabled=False)
         else:
             amp = torch.cpu.amp.autocast(enabled=args.bf16)
-        inputs_channels = data[
-            'paper'].num_features if dataset_name == 'ogbn-mag' \
-            else data.num_features
+        inputs_channels = data.num_features
 
-        for model_name in args.models:
-            if model_name not in supported_sets[dataset_name]:
-                print(f'Configuration of {dataset_name} + {model_name} '
-                      f'not supported. Skipping.')
-                continue
-            print(f'Training bench for {model_name}:')
+        for batch_size in args.batch_sizes:
+            for layers in args.num_layers:
+                num_neighbors = args.num_neighbors
+                if type(num_neighbors) is list:
+                    if len(num_neighbors) == 1:
+                        num_neighbors = num_neighbors * layers
+                elif type(num_neighbors) is int:
+                    num_neighbors = [num_neighbors] * layers
 
-            for batch_size in args.batch_sizes:
-                num_nodes = int(mask[-1].sum()) if hetero else int(mask.sum())
-                sampler = torch.utils.data.RandomSampler(
-                    range(num_nodes), num_samples=args.num_steps *
-                    batch_size) if args.num_steps != -1 else None
-
-                for layers in args.num_layers:
-                    num_neighbors = args.num_neighbors
-                    if type(num_neighbors) is list:
-                        if len(num_neighbors) == 1:
-                            num_neighbors = num_neighbors * layers
-                    elif type(num_neighbors) is int:
-                        num_neighbors = [num_neighbors] * layers
-
-                    assert len(
-                        num_neighbors) == layers, \
-                        f'''num_neighbors={num_neighbors} lenght
+                assert len(
+                    num_neighbors) == layers, \
+                    f'''num_neighbors={num_neighbors} length
                         != num of layers={layers}'''
 
-                    subgraph_loader = NeighborLoader(
-                        data,
-                        num_neighbors=num_neighbors,
-                        input_nodes=mask,
-                        batch_size=batch_size,
-                        shuffle=shuffle,
-                        num_workers=args.num_workers,
-                        sampler=sampler,
-                    )
-                    for hidden_channels in args.num_hidden_channels:
-                        print('----------------------------------------------')
-                        print(f'Batch size={batch_size}, '
-                              f'Layers amount={layers}, '
-                              f'Num_neighbors={num_neighbors}, '
-                              f'Hidden features size={hidden_channels}, '
-                              f'Sparse tensor={args.use_sparse_tensor}')
+                subgraph_loader = NeighborLoader(
+                    data,
+                    num_neighbors=num_neighbors,
+                    batch_size=batch_size,
+                    shuffle=shuffle,
+                    num_workers=args.num_workers,
+                )
 
-                        params = {
-                            'inputs_channels': inputs_channels,
-                            'hidden_channels': hidden_channels,
-                            'output_channels': num_classes,
-                            'num_heads': args.num_heads,
-                            'num_layers': layers,
-                        }
+                for hidden_channels in args.num_hidden_channels:
+                    print('----------------------------------------------')
+                    print(f'Batch size={batch_size}, '
+                          f'Layers amount={layers}, '
+                          f'Num_neighbors={num_neighbors}, '
+                          f'Hidden features size={hidden_channels}, '
+                          f'Sparse tensor={args.use_sparse_tensor}')
 
-                        if model_name == 'pna':
-                            if degree is None:
-                                degree = PNAConv.get_degree_histogram(
-                                    subgraph_loader)
-                                print(f'Calculated degree for {dataset_name}.')
-                            params['degree'] = degree
+                    model = SAGE(inputs_channels,
+                                 hidden_channels, num_classes)
+                    model = model.to(device)
+                    model.train()
+                    optimizer = torch.optim.Adam(model.parameters(),
+                                                 lr=0.001)
 
-                        model = get_model(
-                            model_name, params,
-                            metadata=data.metadata() if hetero else None)
-                        model = model.to(device)
-                        model.train()
-                        optimizer = torch.optim.Adam(model.parameters(),
-                                                     lr=0.001)
+                    progress_bar = False if args.no_progress_bar else True
+                    train = train_homo
+                    with amp:
+                        for _ in range(args.warmup):
+                            train(model, subgraph_loader, optimizer,
+                                  device, progress_bar=progress_bar,
+                                  desc="Warmup")
+                        with timeit(avg_time_divisor=args.num_epochs):
+                            # becomes a no-op if vtune_profile == False
+                            with emit_itt(args.vtune_profile):
+                                for epoch in range(args.num_epochs):
+                                    train(model, subgraph_loader,
+                                          optimizer, device,
+                                          progress_bar=progress_bar,
+                                          desc=f"Epoch={epoch}")
 
-                        progress_bar = False if args.no_progress_bar else True
-                        train = train_hetero if hetero else train_homo
-                        with amp:
-                            for _ in range(args.warmup):
+                        if args.profile:
+                            with torch_profile():
                                 train(model, subgraph_loader, optimizer,
                                       device, progress_bar=progress_bar,
-                                      desc="Warmup")
-                            with timeit(avg_time_divisor=args.num_epochs):
-                                # becomes a no-op if vtune_profile == False
-                                with emit_itt(args.vtune_profile):
-                                    for epoch in range(args.num_epochs):
-                                        train(model, subgraph_loader,
-                                              optimizer, device,
-                                              progress_bar=progress_bar,
-                                              desc=f"Epoch={epoch}")
-
-                            if args.profile:
-                                with torch_profile():
-                                    train(model, subgraph_loader, optimizer,
-                                          device, progress_bar=progress_bar,
-                                          desc="Profile training")
-                                rename_profile_file(model_name, dataset_name,
+                                      desc="Profile training")
+                                rename_profile_file(dataset_name,
                                                     str(batch_size),
                                                     str(layers),
                                                     str(hidden_channels),
@@ -161,7 +117,7 @@ def run(args: argparse.ArgumentParser) -> None:
 if __name__ == '__main__':
     argparser = argparse.ArgumentParser('GNN training benchmark')
     argparser.add_argument('--datasets', nargs='+',
-                           default=['ogbn-mag', 'ogbn-products',
+                           default=['ogbn-products',
                                     'Reddit'], type=str)
     argparser.add_argument(
         '--use-sparse-tensor', action='store_true',
@@ -193,7 +149,8 @@ if __name__ == '__main__':
     argparser.add_argument(
         '--num-steps', default=-1, type=int,
         help='number of steps, -1 means iterating through all the data')
-
+    argparser.add_argument('--gpu_num', type=int, default=2)
+    argparser.add_argument('--micro_pergpu', type=int, default=1)
     args = argparser.parse_args()
 
-    run(args)
+    main(args)