+++ utils/microbatch_reddit_quiver_base.py
@@ -83,7 +83,7 @@ def run(rank, world_size, data, x, quiver_sampler: quiver.pyg.GraphSageSampler, 
                                           shuffle=False, num_workers=6)
 
     torch.manual_seed(12345)
-    model = SAGE(dataset.num_features, 256, dataset.num_classes).to(rank)
+    model = SAGE(data.num_features, 256, dataset.num_classes).to(rank)
     model = DistributedDataParallel(model, device_ids=[rank])
     optimizer = torch.optim.Adam(model.parameters(), lr=0.01)